fix: go_router 부분 app.dart로 분리
